requirements.txt
datasets==3.6.0
numpy==1.26.2
PyYAML==6.0.2
safetensors==0.5.3
tqdm==4.67.1
transformers==4.52.4
triton==3.2.0
wandb==0.21.0
cuda-python==12.9.0